DEL: Remove detritus from earlier VS versions
