+++ Cyotek.AddProjects.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio 15
+VisualStudioVersion = 15.0.27004.2002
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Files", "Solution Files", "{FEA59952-CC14-40A8-85D8-B897601074F8}"
 	ProjectSection(SolutionItems) = preProject
@@ -11,7 +11,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Files", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cyotek.VisualStudioExtensions.AddProjects.vs2015", "src\Cyotek.VisualStudioExtensions.AddProjects.vs2015.csproj", "{A2EBDF4C-FB22-464B-AB34-4ED9EB38F0E1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cyotek.VisualStudioExtensions.AddProjects.2019", "src\Cyotek.VisualStudioExtensions.AddProjects.csproj", "{A2EBDF4C-FB22-464B-AB34-4ED9EB38F0E1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -26,5 +26,8 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BA80B475-96B7-437F-BB0B-0A9249C7A1CB}
 	EndGlobalSection
 EndGlobal
