--- Cyotek.AddProjects.vs2017.sln
@@ -1,33 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2002
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Files", "Solution Files", "{FEA59952-CC14-40A8-85D8-B897601074F8}"
-	ProjectSection(SolutionItems) = preProject
-		CHANGELOG.md = CHANGELOG.md
-		CONTRIBUTORS.md = CONTRIBUTORS.md
-		LICENSE.txt = LICENSE.txt
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cyotek.VisualStudioExtensions.AddProjects.2017", "src\Cyotek.VisualStudioExtensions.AddProjects.2017.csproj", "{A2EBDF4C-FB22-464B-AB34-4ED9EB38F0E1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A2EBDF4C-FB22-464B-AB34-4ED9EB38F0E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A2EBDF4C-FB22-464B-AB34-4ED9EB38F0E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A2EBDF4C-FB22-464B-AB34-4ED9EB38F0E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A2EBDF4C-FB22-464B-AB34-4ED9EB38F0E1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BA80B475-96B7-437F-BB0B-0A9249C7A1CB}
-	EndGlobalSection
-EndGlobal
