--- resources/vs2017/source.extension.vsixmanifest
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
-    <Metadata>
-        <Identity Id="c90738b0-12ac-4ed4-987b-3c57512db4f1" Version="1.0.7.0" Language="en-US" Publisher="Cyotek" />
-        <DisplayName>Add Existing Projects</DisplayName>
-        <Description xml:space="preserve">A simple extension that allows you to easily add multiple projects to a solution with having to manually browse for project files.</Description>
-        <MoreInfo>http://www.cyotek.com/blog/tag/vsix</MoreInfo>
-        <License>Resources\license.txt</License>
-        <Icon>Resources\Package.ico</Icon>
-        <PreviewImage>Resources\preview-image.png</PreviewImage>
-        <Tags />
-    </Metadata>
-    <Installation InstalledByMsi="false">
-        <InstallationTarget Version="[15.0,]" Id="Microsoft.VisualStudio.Community" />
-    </Installation>
-    <Dependencies>
-        <Dependency Id="Microsoft.Framework.NDP" DisplayName="Microsoft .NET Framework" d:Source="Manual" Version="[4.6,)" />
-        <Dependency Id="Microsoft.VisualStudio.MPF.11.0" DisplayName="Visual Studio MPF 11.0" d:Source="Installed" Version="11.0" />
-    </Dependencies>
-    <Assets>
-        <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="%CurrentProject%" Path="|%CurrentProject%;PkgdefProjectOutputGroup|" />
-    </Assets>
-    <Prerequisites>
-        <Prerequisite Id="Microsoft.Net.ComponentGroup.TargetingPacks.Common" Version="[15.0.26606.0,16.0)" DisplayName=".NET Framework 4 – 4.6 development tools" />
-    </Prerequisites>
-</PackageManifest>
