--- src/AddProjectsDialog.xaml
@@ -1,28 +1,0 @@
-﻿<platformUi:DialogWindow x:Class="Cyotek.VisualStudioExtensions.AddProjects.AddProjectsDialog"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                 xmlns:platformUi="clr-namespace:Microsoft.VisualStudio.PlatformUI;assembly=Microsoft.VisualStudio.Shell.11.0"
-                 Title="Add Projects" Height="600" Width="800" ShowInTaskbar="False" WindowStartupLocation="CenterOwner">
-  <Grid Margin="0,0,0,0">
-    <Grid.RowDefinitions>
-      <RowDefinition Height="*"/>
-      <RowDefinition Height="46"/>
-    </Grid.RowDefinitions>
-    <Label x:Name="ProjectsLabel" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="10,10,0,0" Content="_Projects:"/>
-    <ListView x:Name="ProjectsListView" Margin="10,41,10,10">
-      <ListView.View>
-        <GridView>
-          <GridViewColumn Width="30" Header=""/>
-          <GridViewColumn Width ="360" Header="File name"/>
-        </GridView>
-      </ListView.View>
-      <ListBoxItem/>
-      <ListViewItem Content="ListViewItem" Margin="0"/>
-    </ListView>
-    <Button x:Name="AddFileButton" Content="_Add File..." HorizontalAlignment="Left" VerticalAlignment="Bottom" Width="75" Margin="10,0,0,10" Grid.Row="1" Click="AddFileButton_Click"/>
-    <Button x:Name="AddFolderButton" Content="Add _Folder..." HorizontalAlignment="Left" VerticalAlignment="Bottom" Width="75" Margin="90,0,0,10" Grid.Row="1" RenderTransformOrigin="1.858,-0.42" Click="AddFolderButton_Click"/>
-    <Button x:Name="CancelButton" Content="Cancel" HorizontalAlignment="Right" VerticalAlignment="Bottom" Width="75" Margin="0,0,10,10" Grid.Row="1" Click="CancelButton_Click"/>
-    <Button x:Name="OkButton" Content="OK" HorizontalAlignment="Right" VerticalAlignment="Bottom" Width="75" Margin="0,0,90,10" Grid.Row="1" Click="OkButton_Click"/>
-  </Grid>
-
-</platformUi:DialogWindow>
